+++ lib/Braintree/DisputeGateway.php
@@ -69,27 +69,32 @@ class DisputeGateway
      * Adds file evidence to a dispute, given a dispute ID and a document ID
      *
      * @param string $disputeId
-     * @param string $documentId
+     * @param string $documentIdOrRequest
      */
-    public function addFileEvidence($disputeId, $documentId)
+    public function addFileEvidence($disputeId, $documentIdOrRequest)
     {
+        $request = is_array($documentIdOrRequest) ? $documentIdOrRequest : ['documentId' => $documentIdOrRequest];
+
         if (trim($disputeId) == "") {
             throw new Exception\NotFound('dispute with id "' . $disputeId . '" not found');
         }
 
-        if (trim($documentId) == "") {
-            throw new Exception\NotFound('document with id "' . $documentId . '" not found');
+        if (trim($request['documentId']) == "") {
+            throw new Exception\NotFound('document with id "' . $request['documentId'] . '" not found');
         }
 
         try {
-            if (trim($disputeId) == "") {
-                throw new Exception\NotFound();
+            if (array_key_exists('category', $request)) {
+                if (trim($request['category']) == "") {
+                    throw new InvalidArgumentException('category cannot be blank');
+                }
             }
 
+            $request['document_upload_id'] = $request['documentId'];
+            unset($request['documentId']);
+
             $path = $this->_config->merchantPath() . '/disputes/' . $disputeId . '/evidence';
-            $response = $this->_http->post($path, [
-                'document_upload_id' => $documentId
-            ]);
+            $response = $this->_http->post($path, ['evidence' => $request]);
 
             if (isset($response['apiErrorResponse'])) {
                 return new Result\Error($response['apiErrorResponse']);
@@ -127,10 +132,14 @@ class DisputeGateway
             }
 
             if (array_key_exists('tag', $request)) {
-                if (trim($request['tag']) == "") {
-                    throw new InvalidArgumentException('tag cannot be blank');
-                }
                 $evidence['category'] = $request['tag'];
+            }
+
+            if (array_key_exists('category', $request)) {
+                if (trim($request['category']) == "") {
+                    throw new InvalidArgumentException('category cannot be blank');
+                }
+                $evidence['category'] = $request['category'];
             }
 
             if (array_key_exists('sequenceNumber', $request)) {