+++ lib/Braintree/Gateway.php
@@ -256,5 +256,14 @@ class Gateway
     {
         return new WebhookNotificationGateway($this);
     }
+
+    /**
+     *
+     * @return WebhookTestingGateway
+     */
+    public function webhookTesting()
+    {
+        return new WebhookTestingGateway($this);
+    }
 }
 class_alias('Braintree\Gateway', 'Braintree_Gateway');