+++ tests/unit/DisputeTest.php
@@ -29,16 +29,18 @@ class DisputeTest extends Setup
             'reason' => 'fraud',
             'reasonCode' => '83',
             'reasonDescription' => 'Reason code 83 description',
-            'receivedDate' => DateTime::createFromFormat('Ymd', '20130410'),
+            'receivedDate' => DateTime::createFromFormat('Ymd-His', '20130410-000410'),
             'referenceNumber' => '123456',
-            'replyByDate' => DateTime::createFromFormat('Ymd', '20130417'),
+            'replyByDate' => DateTime::createFromFormat('Ymd-His', '20130417-0000417'),
             'status' => 'open',
             'updatedAt' => DateTime::createFromFormat('Ymd-His', '20130410-105039'),
             'evidence' => [[
+                'category' => NULL,
                 'comment' => NULL,
                 'createdAt' => DateTime::createFromFormat('Ymd-His', '20130411-105039'),
                 'id' => 'evidence1',
                 'sentToProcessorAt' => NULL,
+                'sequenceNumber' => NULL,
                 'url' => 'url_of_file_evidence',
             ],[
                 'comment' => 'text evidence',
@@ -74,14 +76,14 @@ class DisputeTest extends Setup
             'currencyIsoCode' => 'USD',
             'status' => 'open',
             'amount' => '100.00',
-            'receivedDate' => DateTime::createFromFormat('Ymd', '20130410'),
-            'replyByDate' => DateTime::createFromFormat('Ymd', '20130410'),
+            'receivedDate' => DateTime::createFromFormat('Ymd-His', '20130410-000410'),
+            'replyByDate' => DateTime::createFromFormat('Ymd-His', '20130421-000421'),
             'reason' => 'fraud',
             'transactionIds' => [
                 'asdf', 'qwer'
             ],
-            'dateOpened' => DateTime::createFromFormat('Ymd', '20130401'),
-            'dateWon' =>DateTime::createFromFormat('Ymd', '20130402'),
+            'dateOpened' => DateTime::createFromFormat('Ymd-His', '20130410-000410'),
+            'dateWon' =>DateTime::createFromFormat('Ymd-His', '20130422-000422'),
             'kind' => 'chargeback'
         ];
 
@@ -95,8 +97,8 @@ class DisputeTest extends Setup
         $this->assertEquals(Braintree\Dispute::Open, $dispute->status);
         $this->assertEquals('transaction_id', $dispute->transactionDetails->id);
         $this->assertEquals('100.00', $dispute->transactionDetails->amount);
-        $this->assertEquals(DateTime::createFromFormat('Ymd', '20130401'), $dispute->dateOpened);
-        $this->assertEquals(DateTime::createFromFormat('Ymd', '20130402'), $dispute->dateWon);
+        $this->assertEquals(DateTime::createFromFormat('Ymd-His', '20130410-000410'), $dispute->dateOpened);
+        $this->assertEquals(DateTime::createFromFormat('Ymd-His', '20130422-000422'), $dispute->dateWon);
         $this->assertEquals(Braintree\Dispute::CHARGEBACK, $dispute->kind);
     }
 
@@ -134,8 +136,10 @@ class DisputeTest extends Setup
         $this->assertEquals(DateTime::createFromFormat('Ymd-His', '20130410-105039'), $dispute->updatedAt);
         $this->assertNull($dispute->evidence[0]->comment);
         $this->assertEquals(DateTime::createFromFormat('Ymd-His', '20130411-105039'), $dispute->evidence[0]->createdAt);
+        $this->assertNull($dispute->evidence[0]->category);
         $this->assertEquals('evidence1', $dispute->evidence[0]->id);
         $this->assertNull($dispute->evidence[0]->sentToProcessorAt);
+        $this->assertNull($dispute->evidence[0]->sequenceNumber);
         $this->assertEquals('url_of_file_evidence', $dispute->evidence[0]->url);
         $this->assertEquals('text evidence', $dispute->evidence[1]->comment);
         $this->assertEquals(DateTime::createFromFormat('Ymd-His', '20130411-105039'), $dispute->evidence[1]->createdAt);
@@ -231,7 +235,7 @@ class DisputeTest extends Setup
         Braintree\Dispute::addTextEvidence("disputeId",
             [
                 'content' => ' ',
-                'tag' => 'CARRIER_NAME',
+                'category' => 'CARRIER_NAME',
                 'sequenceNumber' => '0',
             ]
         );
@@ -244,20 +248,20 @@ class DisputeTest extends Setup
         Braintree\Dispute::addTextEvidence("disputeId",
             [
                 'content' => null,
-                'tag' => 'CARRIER_NAME',
+                'category' => 'CARRIER_NAME',
                 'sequenceNumber' => '0',
             ]
         );
     }
 
-	public function testAddTextEvidenceBlankRequestTagRaisesValueException()
+	public function testAddTextEvidenceBlankRequestCategoryRaisesValueException()
     {
-        $this->setExpectedException('InvalidArgumentException', 'tag cannot be blank');
+        $this->setExpectedException('InvalidArgumentException', 'category cannot be blank');
 
         Braintree\Dispute::addTextEvidence("disputeId",
             [
                 'content' => 'UPS',
-                'tag' => '',
+                'category' => '',
                 'sequenceNumber' => '0',
             ]
         );
@@ -270,7 +274,7 @@ class DisputeTest extends Setup
         Braintree\Dispute::addTextEvidence("disputeId",
             [
                 'content' => 'UPS',
-                'tag' => 'CARRIER_NAME',
+                'category' => 'CARRIER_NAME',
                 'sequenceNumber' => '',
             ]
         );
@@ -283,7 +287,7 @@ class DisputeTest extends Setup
         Braintree\Dispute::addTextEvidence("disputeId",
             [
                 'content' => 'UPS',
-                'tag' => 'CARRIER_NAME',
+                'category' => 'CARRIER_NAME',
                 'sequenceNumber' => '4.5',
             ]
         );
@@ -296,38 +300,74 @@ class DisputeTest extends Setup
         Braintree\Dispute::addTextEvidence("disputeId",
             [
                 'content' => 'UPS',
-                'tag' => 'CARRIER_NAME',
+                'category' => 'CARRIER_NAME',
                 'sequenceNumber' => 'Blah',
             ]
         );
     }
 
-	public function testAddFileEvidenceEmptyIdRaisesNotFoundException()
+    public function testAddFileEvidenceEmptyIdRaisesNotFoundException()
     {
         $this->setExpectedException('Braintree\Exception\NotFound', 'dispute with id " " not found');
 
         Braintree\Dispute::addFileEvidence(" ", 1);
     }
 
-	public function testAddFileEvidenceNullIdRaisesNotFoundException()
+    public function testAddFileEvidenceNullIdRaisesNotFoundException()
     {
         $this->setExpectedException('Braintree\Exception\NotFound', 'dispute with id "" not found');
 
         Braintree\Dispute::addFileEvidence(null, 1);
     }
 
-	public function testAddFileEvidenceEmptyEvidenceRaisesValueException()
+    public function testAddFileEvidenceEmptyEvidenceRaisesValueException()
     {
         $this->setExpectedException('Braintree\Exception\NotFound', 'document with id " " not found');
 
         Braintree\Dispute::addFileEvidence("disputeId", " ");
     }
 
-	public function testAddFileEvidenceNullEvidenceRaisesValueException()
+    public function testAddFileEvidenceNullEvidenceRaisesValueException()
     {
         $this->setExpectedException('Braintree\Exception\NotFound', 'document with id "" not found');
 
         Braintree\Dispute::addFileEvidence("disputeId", null);
+    }
+
+    public function testAddFileEvidenceBlankRequestContentRaisesValueException()
+    {
+        $this->setExpectedException('Braintree\Exception\NotFound', 'document with id " " not found');
+
+        Braintree\Dispute::addFileEvidence("disputeId",
+            [
+                'documentId' => ' ',
+                'category' => 'GENERAL',
+            ]
+        );
+    }
+
+    public function testAddFileEvidenceNullRequestContentRaisesValueException()
+    {
+        $this->setExpectedException('Braintree\Exception\NotFound', 'document with id "" not found');
+
+        Braintree\Dispute::addFileEvidence("disputeId",
+            [
+                'documentId' => null,
+                'category' => 'GENERAL',
+            ]
+        );
+    }
+
+    public function testAddFileEvidenceBlankRequestCategoryRaisesValueException()
+    {
+        $this->setExpectedException('InvalidArgumentException', 'category cannot be blank');
+
+        Braintree\Dispute::addFileEvidence("disputeId",
+            [
+                'documentId' => '123',
+                'category' => '',
+            ]
+        );
     }
 
 	public function testFinalizeNullRaisesNotFoundException()