+++ CHANGELOG.md
@@ -1,8 +1,16 @@
+## Unreleased
+
+## 3.31.0
+* Fix issue where webhook verification would fail due to missing global public key configuration value
+* Fix issue where webhook testing did not work on instantiated gateway
+* Add support for VCR compelling evidence dispute representment
+
 ## 3.30.0
 * Add `oauthAccessRevocation` to `WebhookNotification`s
 * Add support for `profileId` in Transaction#create options for VenmoAccounts
 * Add support for dispute search by `customerId`, `disbursementDate`, and `effectiveDate`
 * Make `CustomerGateway::find` backward compatible
+* Remove `sepaMandateType` and `sepaMandateAcceptanceLocation` params from `ClientTokenGateway`
 
 ## 3.29.0
 * Add support for `association_filter_id` in `Customer#find`